refactor: address-services api call folders
