--- src/services/address-services/addAddressService.js
@@ -1,12 +1,0 @@
-import axios from "axios";
-
-const addAddressService = async (address, encodedToken) =>
-  await axios.post(
-    "/api/user/address",
-    {
-      address: address,
-    },
-    { headers: { authorization: encodedToken } }
-  );
-
-  export default addAddressService;
