--- src/services/address-services/getAddressesService.js
@@ -1,7 +1,0 @@
-import axios from "axios";
-
-const getAddressesService = async(encodedToken) => await axios.get('/api/user/addresses',{
-    headers: { authorization: encodedToken}
-})
-
-export default getAddressesService;
