--- src/services/address-services/removeAddressService.js
@@ -1,8 +1,0 @@
-import axios from "axios";
-
-const removeAddressService = async (addressId, encodedToken) =>
-  await axios.delete(`/api/user/address/${addressId}`, {
-    headers: { authorization: encodedToken },
-  });
-
-export default removeAddressService;
